pyproject.toml
[tool.poetry]
name = "proofreader"
version = "0.1.0"
description = ""
authors = ["Kverde <kverde@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
crispy-bootstrap5 = "^0.6"
django-crispy-forms = "^1.14.0"
whitenoise = "^6.2.0"
gunicorn = "^20.1.0"
psycopg2-binary = "*"
platformdirs = "^2.5.2"
django-tz-detect = "^0.4.0"
django-environ = "^0.10.0"
celery = "^5.3.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.5.0"
django-lifecycle = "^1.0.0"
pypdf2 = "^3.0.1"
pdf2image = "^1.16.3"
pytesseract = "^0.3.10"
django-extensions = "^3.2.3"
ipython = "^8.14.0"
django-simple-history = "^3.3.0"
sentry-sdk = "^1.32.0"


[tool.poetry.dev-dependencies]
black = "^22.6.0"
djlint = "^1.11.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
(
  /(
      \.git
    | \.mypy_cache
    | \.venv
    | \.tox
    | build
    | .*/migrations
  )/
)
'''
skip-string-normalization = true

# settings are moved from setup.cfg since seed-isort-config populate only
# .isort.cfg and pyproject.toml files and hence
# pyproject.toml is used to eliminate extra .isort.cfg file
[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
indent = "    "
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
no_lines_before = "LOCALFOLDER"
skip_glob = ["**/migrations/**"]
include_trailing_comma = true
known_third_party=[]